% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toenail}
\alias{toenail}
\title{Name_of_the_data}
\format{A data frame with 100 rows and 2 variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}}
\usage{
toenail
}
\description{
Description_of_the_Data
}
\keyword{datasets}
