% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLTranktest.R
\name{LLTranktest}
\alias{LLTranktest}
\title{LLT rank-based non-parameteric test}
\usage{
LLTranktest(formula, id = NULL, data, B = 200)
}
\arguments{
\item{formula, }{id=NULL, data, B=118}
}
\value{
data.frame
}
\description{
LLT rank-based non-parameteric test
}
\examples{
LLTranktest(Y~X, id=mydata$ID, data=mydata)

}
